+++ google-cloud-bigquerystorage/src/main/java/com/google/cloud/bigquery/storage/v1/ConnectionWorkerPool.java
@@ -27,6 +27,7 @@ import com.google.common.base.Preconditions;
 import com.google.common.base.Stopwatch;
 import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.MoreExecutors;
+import io.opentelemetry.api.common.Attributes;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.Comparator;
@@ -233,14 +234,7 @@ public class ConnectionWorkerPool {
     ConnectionWorkerPool.settings = settings;
   }
 
-  /** Distributes the writing of a message to an underlying connection. */
-  ApiFuture<AppendRowsResponse> append(StreamWriter streamWriter, ProtoRows rows) {
-    return append(streamWriter, rows, -1);
-  }
-
-  /** Distributes the writing of a message to an underlying connection. */
-  ApiFuture<AppendRowsResponse> append(StreamWriter streamWriter, ProtoRows rows, long offset) {
-    // We are in multiplexing mode after entering the following logic.
+  ConnectionWorker getConnectionWorker(StreamWriter streamWriter) {
     ConnectionWorker connectionWorker;
     lock.lock();
     try {
@@ -277,9 +271,17 @@ public class ConnectionWorkerPool {
     } finally {
       lock.unlock();
     }
+    return connectionWorker;
+  }
+
+  /** Distributes the writing of a message to an underlying connection. */
+  ApiFuture<AppendRowsResponse> append(
+      StreamWriter streamWriter, ProtoRows rows, long offset, String uniqueRequestId) {
+    // We are in multiplexing mode after entering the following logic.
+    ConnectionWorker connectionWorker = getConnectionWorker(streamWriter);
     Stopwatch stopwatch = Stopwatch.createStarted();
     ApiFuture<AppendRowsResponse> responseFuture =
-        connectionWorker.append(streamWriter, rows, offset);
+        connectionWorker.append(streamWriter, rows, offset, uniqueRequestId);
     return ApiFutures.transform(
         responseFuture,
         // Add callback for update schema
@@ -292,6 +294,12 @@ public class ConnectionWorkerPool {
           return response;
         },
         MoreExecutors.directExecutor());
+  }
+
+  @VisibleForTesting
+  Attributes getTelemetryAttributes(StreamWriter streamWriter) {
+    ConnectionWorker connectionWorker = getConnectionWorker(streamWriter);
+    return connectionWorker.getTelemetryAttributes();
   }
 
   /**