+++ google-cloud-bigquerystorage/src/test/java/com/google/cloud/bigquery/storage/v1/StreamWriterTest.java
@@ -56,6 +56,7 @@ import com.google.protobuf.Descriptors.DescriptorValidationException;
 import com.google.protobuf.Int64Value;
 import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
+import io.opentelemetry.api.common.Attributes;
 import java.io.IOException;
 import java.time.Instant;
 import java.util.ArrayList;
@@ -165,6 +166,8 @@ public class StreamWriterTest {
         .setLocation("US")
         .setEnableConnectionPool(true)
         .setMaxRetryDuration(java.time.Duration.ofSeconds(5))
+        // We won't test profiler behavior in this test, it's a sanity check.
+        .setEnableLatencyProfiler(true)
         .build();
   }
 
@@ -819,6 +822,42 @@ public class StreamWriterTest {
     assertTrue(writer2.getInflightWaitSeconds() >= 1);
     assertEquals(0, appendFuture1.get().getAppendResult().getOffset().getValue());
     assertEquals(1, appendFuture2.get().getAppendResult().getOffset().getValue());
+    writer1.close();
+    writer2.close();
+  }
+
+  @Test
+  public void testOpenTelemetryAttributes_MultiplexingCase() throws Exception {
+    ConnectionWorkerPool.setOptions(
+        Settings.builder().setMinConnectionsPerRegion(1).setMaxConnectionsPerRegion(1).build());
+    StreamWriter writer1 =
+        StreamWriter.newBuilder(TEST_STREAM_1, client)
+            .setWriterSchema(createProtoSchema())
+            .setLocation("US")
+            .setEnableConnectionPool(true)
+            .build();
+    StreamWriter writer2 =
+        StreamWriter.newBuilder(TEST_STREAM_2, client)
+            .setWriterSchema(createProtoSchema())
+            .setLocation("US")
+            .setEnableConnectionPool(true)
+            .build();
+
+    testBigQueryWrite.addResponse(createAppendResponse(0));
+    testBigQueryWrite.addResponse(createAppendResponse(1));
+
+    ApiFuture<AppendRowsResponse> appendFuture1 = sendTestMessage(writer1, new String[] {"A"});
+    assertEquals(0, appendFuture1.get().getAppendResult().getOffset().getValue());
+    Attributes attributes = writer1.getTelemetryAttributes();
+    String attributesTableId = attributes.get(ConnectionWorker.telemetryKeyTableId);
+    assertEquals("projects/p/datasets/d1/tables/t1", attributesTableId);
+
+    ApiFuture<AppendRowsResponse> appendFuture2 = sendTestMessage(writer2, new String[] {"A"});
+    assertEquals(1, appendFuture2.get().getAppendResult().getOffset().getValue());
+    attributes = writer2.getTelemetryAttributes();
+    attributesTableId = attributes.get(ConnectionWorker.telemetryKeyTableId);
+    assertEquals("projects/p/datasets/d2/tables/t2", attributesTableId);
+
     writer1.close();
     writer2.close();
   }
@@ -1808,6 +1847,35 @@ public class StreamWriterTest {
 
     writer.close();
   }
+
+  /* temporarily disable test as static variable is interfering with other tests
+  @Test
+  public void testInternalQuotaError_MaxWaitTimeExceed_RetrySuccess() throws Exception {
+    // In order for the test to succeed, the given request must complete successfully even after all
+    // the retries. The fake server is configured to fail 3 times with a quota error. This means the
+    // client will perform retry with exponential backoff. The fake server injects 1 second of delay
+    // for each response. In addition, the exponential backoff injects a couple of seconds of delay.
+    // This yields an overall delay of about 5 seconds before the request succeeds. If the request
+    // send timestamp was being set only once, this would eventually exceed the 4 second timeout
+    // limit, and throw an exception. With the current behavior, the request send timestamp is reset
+    // each time a retry is performed, so we never exceed the 4 second timeout limit.
+    StreamWriter.setMaxRequestCallbackWaitTime(java.time.Duration.ofSeconds(4));
+    testBigQueryWrite.setResponseSleep(Duration.ofSeconds(1));
+    StreamWriter writer = getTestStreamWriterRetryEnabled();
+    testBigQueryWrite.addStatusException(
+        com.google.rpc.Status.newBuilder().setCode(Code.RESOURCE_EXHAUSTED.ordinal()).build());
+    testBigQueryWrite.addStatusException(
+        com.google.rpc.Status.newBuilder().setCode(Code.RESOURCE_EXHAUSTED.ordinal()).build());
+    testBigQueryWrite.addStatusException(
+        com.google.rpc.Status.newBuilder().setCode(Code.RESOURCE_EXHAUSTED.ordinal()).build());
+    testBigQueryWrite.addResponse(createAppendResponse(0));
+
+    ApiFuture<AppendRowsResponse> appendFuture1 =
+        writer.append(createProtoRows(new String[] {"A"}));
+    assertEquals(0, appendFuture1.get().getAppendResult().getOffset().getValue());
+    writer.close();
+  }
+  */
 
   @Test
   public void testAppendSuccessAndInternalErrorRetrySuccessExclusive() throws Exception {